+++ pkg/client/generic/scheme/register.go
@@ -23,7 +23,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	k8sscheme "k8s.io/client-go/kubernetes/scheme"
-	crapis "k8s.io/cluster-registry/pkg/apis/clusterregistry/v1alpha1"
 
 	fedapis "github.com/kubernetes-sigs/federation-v2/pkg/apis"
 )
@@ -34,7 +33,6 @@ var ParameterCodec = runtime.NewParameterCodec(Scheme)
 var localSchemeBuilder = runtime.SchemeBuilder{
 	fedapis.AddToScheme,
 	k8sscheme.AddToScheme,
-	crapis.AddToScheme,
 }
 
 func init() {