+++ pkg/controller/federatedcluster/controller.go
@@ -67,9 +67,6 @@ type ClusterController struct {
 	// fedNamespace is the name of the namespace containing
 	// FederatedCluster resources and their associated secrets.
 	fedNamespace string
-
-	// clusterNamespace is the namespace containing Cluster resources.
-	clusterNamespace string
 }
 
 // StartClusterController starts a new cluster controller.
@@ -94,7 +91,6 @@ func newClusterController(config *util.ControllerConfig, clusterHealthCheckConfi
 		clusterHealthCheckConfig: clusterHealthCheckConfig,
 		clusterDataMap:           make(map[string]*ClusterData),
 		fedNamespace:             config.FederationNamespace,
-		clusterNamespace:         config.ClusterNamespace,
 	}
 	var err error
 	_, cc.clusterController, err = util.NewGenericInformerWithEventHandler(
@@ -131,7 +127,7 @@ func (cc *ClusterController) addToClusterSet(obj interface{}) {
 	klog.V(1).Infof("ClusterController observed a new cluster: %v", cluster.Name)
 	// create the restclient of cluster
 	clientTimeout := time.Duration(cc.clusterHealthCheckConfig.TimeoutSeconds) * time.Second
-	restClient, err := NewClusterClientSet(cluster, cc.client, cc.fedNamespace, cc.clusterNamespace, clientTimeout)
+	restClient, err := NewClusterClientSet(cluster, cc.client, cc.fedNamespace, clientTimeout)
 	if err != nil || restClient == nil {
 		klog.Errorf("Failed to create corresponding restclient of kubernetes cluster: %v", err)
 		return