+++ test/e2e/framework/test_context.go
@@ -32,7 +32,6 @@ type TestContextType struct {
 	KubeConfig                      string
 	KubeContext                     string
 	FederationSystemNamespace       string
-	ClusterNamespace                string
 	SingleCallTimeout               time.Duration
 	LimitedScope                    bool
 	LimitedScopeInMemoryControllers bool
@@ -54,8 +53,6 @@ func registerFlags(t *TestContextType) {
 		"kubeconfig context to use/override. If unset, will use value from 'current-context'.")
 	flag.StringVar(&t.FederationSystemNamespace, "federation-namespace", util.DefaultFederationSystemNamespace,
 		fmt.Sprintf("The namespace the federation control plane is deployed in.  If unset, will default to %q.", util.DefaultFederationSystemNamespace))
-	flag.StringVar(&t.ClusterNamespace, "registry-namespace", util.MulticlusterPublicNamespace,
-		fmt.Sprintf("The cluster registry namespace.  If unset, will default to %q.", util.MulticlusterPublicNamespace))
 	flag.DurationVar(&t.SingleCallTimeout, "single-call-timeout", DefaultSingleCallTimeout,
 		fmt.Sprintf("The maximum duration of a single call.  If unset, will default to %v", DefaultSingleCallTimeout))
 	flag.BoolVar(&t.LimitedScope, "limited-scope", false, "Whether the federation namespace (configurable via --federation-namespace) will be the only target for federation.")